Function to create a sentence from individual words

Setting up an R package is easy when using the RStudio GUI as as well as the R packages devtools, usethis, Rdpack, testhat, and roxygen. You can nicely document your functions, and easily share argument definitions across multiple functions with the roxygen @template tag. We can also make our documentation fancy by adding math, for example, like this: \(\sum_{i}^{n} x^i\).

Usage

createSentence(subject, adverb, verb, noun)

Arguments

subject

The subject of the sentence. Choices are limited to 'I', 'You', 'She', 'He', 'They', and 'We'.

adverb

An optional word to modify the verb of the sentence.

verb

The word indicating what action is happening in the sentence.

noun

The word indicating the object to which the action is happening.

Value

Returns a complete sentence as a single string.

References

Wickham H (2015). R packages: organize, test, document, and share your code. " O'Reilly Media, Inc.".

Examples

subject = "I"
adverb = "happily"
verb = "walked"
noun = "home"
createSentence(subject, adverb, verb, noun)
#> [1] "I happily walked home."